fw-status: Fix crash due to bad variant access
After MCTP moved to code construct interfaces, EID's property has
changed from uint32_t to uint8_t. This was causing the app to crash due
to bad_variant_access

Tests:
- FW Status no longer crashes on hgxb300

Fixes NVBug https://nvbugspro.nvidia.com/bug/5568505

+++ fw-status/cx8_resource.hpp
@@ -172,7 +172,7 @@ class Cx8Resource : public MCTPDiscoveryResource
                     continue;
                 }
 
-                const auto& mctpEID = std::get<uint32_t>(
+                const auto& mctpEID = std::get<uint8_t>(
                     interfaces.at(mctpEndpointIntfName).at("EID"));
 
                 if (mctpEID != smaEid)
