+++ fw-status/gpu_resource.hpp
@@ -181,7 +181,7 @@ class GpuResource : public MCTPDiscoveryResource
                     continue;
                 }
 
-                const auto& mctpEID = std::get<uint32_t>(
+                const auto& mctpEID = std::get<uint8_t>(
                     interfaces.at(mctpEndpointIntfName).at("EID"));
 
                 if (mctpEID != smaEid)
